feat: restructure the `utils` module
