+++ colpali_engine/trainer/__init__.py
@@ -1,0 +1,2 @@
+from contrastive_trainer import *
+from retrieval_evaluator import *
