+++ colpali_engine/utils/processing_utils/colidefics_processing_utils.py
@@ -1,7 +1,7 @@
 """Utils for processing images and queries for ColPaLi"""
 
 
-def process_images(processor, images, max_length: int = 50):
+def process_images(processor, images):
     texts_doc = []
     images = [image.convert("RGB") for image in images]
