+++ colpali_engine/utils/processing_utils/colpali_processing_utils.py
